refactor: split device-payload appliers and monolithic platform modules
- _coordinator_devices.py (776 → 328 lines): the 640-line reconciliation
  method now delegates per-family attribute application to _device_apply.py
  — pure (device, device_data) functions grouped by device family, run in
  the original order, honouring the per-key optimistic guards.
- binary_sensor.py / select.py / number.py follow the established house
  pattern: thin platform module (setup + discovery builders, static setup
  folded onto the builders) + _*_entities.py for classes and descriptor
  tables, with re-exports for backwards compatibility. The LightSwitch
  touch-sensitivity definition, previously duplicated inline twice, is
  named once (LIGHTSWITCH_TOUCH_SENSITIVITY_NUMBER).
- config_flow.py: single _build_api helper replaces the four duplicated
  AjaxRestApi constructions (user/proxy/reauth/reconfigure).
